+++ backends/gdx-backends-gwt/src/com/badlogic/gdx/backends/gwt/emu/java/nio/DirectReadOnlyShortBufferAdapter.java
@@ -76,7 +76,7 @@ final class DirectReadOnlyShortBufferAdapter extends ShortBuffer implements HasA
 // if (position == limit) {
 // throw new BufferUnderflowException();
 // }
-        return (short) shortArray.get(position++);
+        return shortArray.get(position++);
     }
 
     @Override
@@ -84,7 +84,7 @@ final class DirectReadOnlyShortBufferAdapter extends ShortBuffer implements HasA
 // if (index < 0 || index >= limit) {
 // throw new IndexOutOfBoundsException();
 // }
-        return (short) shortArray.get(index);
+        return shortArray.get(index);
     }
 
     @Override
