+++ gdx/src/com/badlogic/gdx/utils/UBJsonReader.java
@@ -221,7 +221,7 @@ public class UBJsonReader implements BaseJsonReader{
     throws IOException{
         if(type == 'i') return (long) readUChar(din);
         if(type == 'I') return (long) readUShort(din);
-        if(type == 'l') return (long) readUInt(din);
+        if(type == 'l') return readUInt(din);
         if(type == 'L') return din.readLong();
         if(useIntOnError){
             long result = (long) ((short) type & 0xFF) << 24;
