+++ gdx/src/com/badlogic/gdx/utils/compression/lzma/Encoder.java
@@ -214,8 +214,6 @@ public class Encoder{
         }
     }
 
-    ;
-
     public static final int kNumLenSpecSymbols = Base.kNumLowLenSymbols + Base.kNumMidLenSymbols;
 
     class LenPriceTableEncoder extends LenEncoder{
@@ -275,7 +273,6 @@ public class Encoder{
 
         public void MakeAsShortRep(){
             BackPrev = 0;
-            ;
             Prev1IsChar = false;
         }
 
@@ -283,8 +280,6 @@ public class Encoder{
             return (BackPrev == 0);
         }
     }
-
-    ;
 
     Optimal[] _optimum = new Optimal[kNumOpts];
     com.badlogic.gdx.utils.compression.lz.BinTree _matchFinder = null;
@@ -405,7 +400,7 @@ public class Encoder{
         if(_numDistancePairs > 0){
             lenRes = _matchDistances[_numDistancePairs - 2];
             if(lenRes == _numFastBytes)
-                lenRes += _matchFinder.GetMatchLen((int) lenRes - 1, _matchDistances[_numDistancePairs - 1], Base.kMatchMaxLen
+                lenRes += _matchFinder.GetMatchLen(lenRes - 1, _matchDistances[_numDistancePairs - 1], Base.kMatchMaxLen
                 - lenRes);
         }
         _additionalOffset++;
@@ -990,10 +985,10 @@ public class Encoder{
             int complexState = (_state << Base.kNumPosStatesBitsMax) + posState;
             if(len == 1 && pos == -1){
                 _rangeEncoder.Encode(_isMatch, complexState, 0);
-                byte curByte = _matchFinder.GetIndexByte((int) (0 - _additionalOffset));
+                byte curByte = _matchFinder.GetIndexByte((0 - _additionalOffset));
                 LiteralEncoder.Encoder2 subCoder = _literalEncoder.GetSubCoder((int) nowPos64, _previousByte);
                 if(!Base.StateIsCharState(_state)){
-                    byte matchByte = _matchFinder.GetIndexByte((int) (0 - _repDistances[0] - 1 - _additionalOffset));
+                    byte matchByte = _matchFinder.GetIndexByte((0 - _repDistances[0] - 1 - _additionalOffset));
                     subCoder.EncodeMatched(_rangeEncoder, matchByte, curByte);
                 }else
                     subCoder.Encode(_rangeEncoder, curByte);
@@ -1040,7 +1035,7 @@ public class Encoder{
                     _posSlotEncoder[lenToPosState].Encode(_rangeEncoder, posSlot);
 
                     if(posSlot >= Base.kStartPosModelIndex){
-                        int footerBits = (int) ((posSlot >> 1) - 1);
+                        int footerBits = ((posSlot >> 1) - 1);
                         int baseVal = ((2 | (posSlot & 1)) << footerBits);
                         int posReduced = pos - baseVal;
 
@@ -1161,7 +1156,7 @@ public class Encoder{
     void FillDistancesPrices(){
         for(int i = Base.kStartPosModelIndex; i < Base.kNumFullDistances; i++){
             int posSlot = GetPosSlot(i);
-            int footerBits = (int) ((posSlot >> 1) - 1);
+            int footerBits = ((posSlot >> 1) - 1);
             int baseVal = ((2 | (posSlot & 1)) << footerBits);
             tempPrices[i] = BitTreeEncoder.ReverseGetPrice(_posEncoders, baseVal - posSlot - 1, footerBits, i - baseVal);
         }
